requirements.txt
alabaster==1.0.0
appnope==0.1.4
asttokens==2.4.1
babel==2.16.0
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
cycler==0.12.1
decorator==5.1.1
docutils==0.21.2
executing==2.1.0
fonttools==4.54.1
idna==3.10
imagesize==1.4.1
iniconfig==1.1.1
ipython==8.12.3
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.5.3
matplotlib-inline==0.1.7
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyMySQL==1.0.3
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.14.1
seaborn==0.11.2
setuptools==68.2.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13

websocket-client~=1.8.0